add:test folder

+++ uk-hacktiv8-prakerja/repositories/postgres/user_repository.go
@@ -2,6 +2,7 @@ package postgres
 
 
 import (
+	"fmt"
 	"uk-hacktiv8-prakerja/database"
 	"uk-hacktiv8-prakerja/models"
 )
@@ -16,7 +17,7 @@ func(u *UserRepository)Store(usr models.User)(models.User,error){
 	err := db.Create(&usr).Error
 
 	if err != nil {
-		return modelss.User{}, err
+		return models.User{}, err
 	}
 
 	return usr, nil
@@ -24,26 +25,35 @@ func(u *UserRepository)Store(usr models.User)(models.User,error){
 
 func (s *UserRepository) Delete(id uint) (models.User, error) {
 	db := database.GetDatabase()
-	n, err := s.Get(id)
-	if err != nil {
-		return models.User{}, err
+	var n models.User
+	if !s.IsExist(id){
+		return models.User{},nil
+	}else{
+		n, err := s.Get(id)
+		if err != nil {
+			return models.User{}, err
+		}
+
+		err = db.Delete(&n).Error
+		if err != nil {
+			return models.User{}, fmt.Errorf("tidak bisa menghapus data: %v", err)
+		}
+
 	}
 
-	err = db.Delete(&n).Error
-	if err != nil {
-		return models.User{}, fmt.Errorf("cannot delete file: %v", err)
-	}
 
 	return n, nil
 }
 
 func (s *UserRepository) Update(p models.User) error {
 	db := database.GetDatabase()
-
-	err := db.Save(&p).Error
-	if err != nil {
-		return fmt.Errorf("cannot update client on pg: %v", err)
+	if s.IsExist(p.ID){
+		err := db.Save(&p).Error
+		if err != nil {
+			return fmt.Errorf("tidak bisa mengupdate user di database: %v", err)
+		}
 	}
+
 
 	return nil
 }
@@ -54,7 +64,7 @@ func (s *UserRepository) Get(id uint) (models.User, error) {
 	err := db.First(&p, id).Error
 
 	if err != nil {
-		return models.User{}, fmt.Errorf("cannot find User by id: %v", err)
+		return models.User{}, fmt.Errorf("tidak dapat menemukan user dengan  id: %v", err)
 	}
 
 	return p, nil
@@ -72,4 +82,15 @@ func (s *UserRepository) GetAll() ([]models.User, error) {
 	return p, err
 }
 
+func (s *UserRepository)IsExist(id uint)(bool){
+	db := database.GetDatabase()
+	user := &models.User{
+		ID:id,
+	}
+
+	 result := db.First(user)
+
+	 return result.RowsAffected > 0
+
+}
